requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
clarifai-grpc==8.0.0
cryptography==37.0.2
dnspython==2.2.1
emoji==1.7.0
ephem==4.1.3
flake8==4.0.1
googleapis-common-protos==1.56.2
grpcio==1.46.3
h11==0.13.0
idna==3.3
mccabe==0.6.1
outcome==1.1.0
protobuf==3.20.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pymongo==4.1.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wsproto==1.1.0
